requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
sendgrid==6.11.0
twilio==8.10.0
pandas==2.1.4
jinja2==3.1.2
qrcode==7.4.2
pillow==10.1.0
werkzeug==3.0.1
cryptography==41.0.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
